+++ src/main/cpp/communication.cpp
@@ -427,7 +427,7 @@ int get_local_rank() {
 
 void mpi_startup ( int argc, char* argv[], int block_dim_size ) {
   // can fit 5 blocks of double
-  max_buffer_size = max(100000LU,5*sizeof(double)*block_dim_size*block_dim_size);
+  max_buffer_size = max(100000LU,5*sizeof(float)*block_dim_size*block_dim_size);
   if (inMemory)
     return;
   // Multiple threads may call MPI, with no restrictions
